Add files via upload
This dice game project is a simple yet engaging game for two players, built using HTML, CSS, and JavaScript. Players take turns clicking the "Roll" button to generate random numbers, which are then displayed on the screen. The player with the higher number wins the round, and the game can be played multiple times to determine the overall winner. With its straightforward gameplay and random number generation, this project showcases a fun and interactive application of web development technologies.